requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
openai>=1.30.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
websockets>=12.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langgraph>=0.2.0
